+++ src/macros.rs
@@ -208,7 +208,11 @@ macro_rules! libc_enum {
 /// offset of `field` within struct `ty`
 #[cfg(not(target_os = "redox"))]
 macro_rules! offset_of {
-    ($ty:ty, $field:ident) => {
-        &(*(ptr::null() as *const $ty)).$field as *const _ as usize
-    }
+    ($ty:ty, $field:ident) => {{
+        // Safe because we don't actually read from the dereferenced pointer
+        #[allow(unused_unsafe)] // for when the macro is used in an unsafe block
+        unsafe {
+            &(*(ptr::null() as *const $ty)).$field as *const _ as usize
+        }
+    }}
 }